+++ client/package.json
@@ -18,7 +18,7 @@
     "axios": "^1.8.4",
     "chroma-js": "^3.1.2",
     "react": "^19.0.0",
-    "react-dom": "^19.1.0",
+    "react-dom": "^19.0.0",
     "react-router-dom": "^7.5.0"
   },
   "devDependencies": {
