--- client/src/components/Contact.jsx
@@ -1,9 +1,0 @@
-function Contact() {
-  return (
-    <div style={{ textAlign: "center", marginTop: "40px" }}>
-      <h2>Contact Us</h2>
-    </div>
-  );
-}
-
-export default Contact;
