--- client/src/components/Launch.jsx
@@ -1,9 +1,0 @@
-function Launch() {
-  return (
-    <div style={{ textAlign: "center", marginTop: "40px" }}>
-      <h2>Launch Page</h2>
-    </div>
-  );
-}
-
-export default Launch;
