--- client/src/components/MaintenanceTracker.jsx
@@ -1,9 +1,0 @@
-function MaintenanceTracker() {
-  return (
-    <div style={{ textAlign: "center", marginTop: "40px" }}>
-      <h2>Maintenance Tracker Page</h2>
-    </div>
-  );
-}
-
-export default MaintenanceTracker;
