--- client/src/components/MyRequests.jsx
@@ -1,9 +1,0 @@
-function MyRequests() {
-  return (
-    <div style={{ textAlign: "center", marginTop: "40px" }}>
-      <h2>My Requests Page</h2>
-    </div>
-  );
-}
-
-export default MyRequests;
