+++ client/src/styles/index.css
@@ -1,3 +1,10 @@
+/* color scheme:
+#101420 Dark blue/black - div backgrounds
+#426f4d Dark green - font over background color, mouse over buttons
+#8cc5a6 Light green - button background, div outlines
+#961e14 Subdued red - font over background color, errors
+*/
+
 :root {
   font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -28,7 +35,7 @@ body {
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
-  background-image: url("../public/topomapbackground.png");
+  background-image: url("../topomapbackground.png");
 }
 
 h1 {
